backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==1.4.53
databases[sqlite]==0.7.0
aiosqlite==0.19.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 